refactor(engine): extract spill/stats.rs (SPL-6 #2328)
